+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased
+### Added
+* Added `OutputAssignments`, `OutputAssignment`, `ChannelAssignment`, and `ChanelAssignmentTexture`.
+* Added method `output_assignments` to `Material` for accessing shader JSON database information.
+
+### Changed
+* Changed the types of `MaterialParameters` fields to match xc3_model changes.
+* Renamed `Material.unk_type` to `Material.pass_type` to match xc3_model.
+
+### Removed
+* Removed `BufferParameter`
+
 ## 0.3.0 - 2024-02-28
 ### Added
 * Added support for accessing Rust logs from Python's `logging` module.